+++ R/is_arf.R
@@ -1,0 +1,6 @@
+#' Checks if running R via Alternative R Frontend (arf)
+#'
+#' @return A logical
+is_arf <- function() {
+  basename(commandArgs()[1]) == "arf"
+}